Fixed some bugs, full expression codegen in place.
Code generation for the range of expressions is in place! It's really messy and not optimal in the
slightest, but it seems to work so far!
Found a bug in the lexer - it couldn't identify a division sign in an expression. Fixed that.
Found a few more bugs due to my quick copy pasting. Fixed those as well.