Fixed procedure nesting.
Multiple issues with my old method: I set up the activation record after the procedure call so I had
to fix the static pointer in the generateProcedure method. That was incorrect - I needed to
calculate the link before jumping to the procedure label. That logic is moved now to the Call
statement...
I think the compiler is correct now though!